Finish simple support for Archetype adding and removal.
We only support adding archetypes that are already in the combobox (and therefore
compiled in), next we would need to detect between edited and selected combobox only
to figure out if we should parse the inputted archetype. We need that for newly
added archetypes that haven't yet made it into an update for the editor.