Add custom archetype adding support
This will help for those periods when new archetypes have been announced but
have not yet made it into an update to the editor. It tries to parse the
combobox's text first in decimal and then in hexadecimal, if it succeeds the
actual value is appended to the list and its true value is shown in hexadecimal.