Add support for copy-pasting cards between databases
The way it works is that selection of cards is copied to a in-memory database,
which acts as a "clipboard". It serves as staging area to view and edit changes
before putting them back onto a regular database. This database can also be
visualized in a special tab which can be opened from the File menu.